pyproject.toml
[project]
name = "prompteer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.4",
    "google-auth==2.40.3",
    "google-genai==1.25.0",
    "greenlet==3.2.3",
    "gunicorn>=23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "pillow==11.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt>=2.10.1",
    "pytest-mock>=3.14.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "rich==14.0.0",
    "rich-toolkit==0.14.8",
    "rignore==0.6.2",
    "rsa==4.9.1",
    "sentry-sdk==2.33.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.47.1",
    "tenacity==8.5.0",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]

[tool.pytest.ini_options]
markers = [
    "gemini_api: marks tests that call the live Gemini API",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
start = "fastapi_cli:start"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "vulture>=2.14",
]
